requirements.txt
fastapi
uvicorn
python-multipart
pytesseract
pillow
PyPDF2
pdf2image
python-dotenv
groq
pydantic
slowapi
